+++ autofit/interpolator/__init__.py
@@ -1,0 +1,3 @@
+from .spline import SplineInterpolator
+from .linear import LinearInterpolator
+from .covariance import CovarianceInterpolator
