+++ autofit/interpolator/abstract.py
@@ -1,86 +1,11 @@
 import copy
 from abc import ABC, abstractmethod
-
-import numpy as np
-import scipy
-from scipy.interpolate import CubicSpline
 from typing import List, Dict, cast
-
-from scipy.stats import stats
-from scipy.optimize import curve_fit
 
 from autoconf.dictable import Dictable
 from autofit.mapper.model import ModelInstance
-from autofit.non_linear.samples.pdf import SamplesPDF
 
-
-class InterpolatorPath:
-    def __init__(self, keys: List[str]):
-        """
-        Addresses a given attribute in a ModelInstance
-
-        Parameters
-        ----------
-        keys
-            A list of attribute names
-        """
-        self.keys = keys
-
-    def __getattr__(self, item: str) -> "InterpolatorPath":
-        """
-        Add a new attribute name to the end of the path
-        """
-        return InterpolatorPath(self.keys + [item])
-
-    def get_value(self, instance: ModelInstance) -> float:
-        """
-        Retrieve the value at the path for a given instance.
-
-        Parameters
-        ----------
-        instance
-            An instance of some model
-
-        Returns
-        -------
-        The value for the instance at the given path.
-        """
-        for key in self.keys:
-            instance = getattr(instance, key)
-        return instance
-
-    def __eq__(self, other: float) -> "Equality":
-        """
-        Create an object describing the value the addressed attribute
-        should have for interpolation.
-
-        Parameters
-        ----------
-        other
-            A value to which the instance will be interpolated
-
-        Returns
-        -------
-        An object describing how to interpolate
-        """
-        return Equality(self, other)
-
-
-class Equality:
-    def __init__(self, path: InterpolatorPath, value: float):
-        """
-        Describes the value of a given attribute for which other values
-        are interpolated.
-
-        Parameters
-        ----------
-        path
-            The path to an attribute
-        value
-            The value that attribute should have
-        """
-        self.path = path
-        self.value = value
+from .query import InterpolatorPath, Equality
 
 
 class AbstractInterpolator(Dictable, ABC):
@@ -204,77 +129,3 @@ class AbstractInterpolator(Dictable, ABC):
         -------
         An interpolated value for the attribute
         """
-
-
-class LinearInterpolator(AbstractInterpolator):
-    """
-    Assume all attributes have a linear relationship with time
-    """
-
-    @staticmethod
-    def _interpolate(x, y, value):
-        slope, intercept, r, p, std_err = stats.linregress(x, y)
-        return slope * value + intercept
-
-
-class SplineInterpolator(AbstractInterpolator):
-    """
-    Interpolate data with a piecewise cubic polynomial which is twice continuously differentiable
-    """
-
-    @staticmethod
-    def _interpolate(x, y, value):
-        f = CubicSpline(x, y)
-        return f(value)
-
-
-class CovarianceInterpolator(AbstractInterpolator):
-    def __init__(
-        self,
-        samples_list: List[SamplesPDF],
-    ):
-        self.samples_list = samples_list
-        # noinspection PyTypeChecker
-        super().__init__([samples.max_log_likelihood() for samples in samples_list])
-
-    @property
-    def covariance_matrix(self):
-        matrices = [samples.covariance_matrix() for samples in self.samples_list]
-        prior_count = self.samples_list[0].model.prior_count
-        size = prior_count * len(self.samples_list)
-        array = np.zeros((size, size))
-        for i, matrix in enumerate(matrices):
-            array[
-                i * prior_count : (i + 1) * prior_count,
-                i * prior_count : (i + 1) * prior_count,
-            ] = matrix
-        return array
-
-    @property
-    def inverse_covariance_matrix(self):
-        return scipy.linalg.inv(self.covariance_matrix)
-
-    @staticmethod
-    def _interpolate(x, y, value):
-        pass
-
-    def __getitem__(self, value: Equality):
-        x = [value.path.get_value(instance) for instance in self.instances]
-
-        def func(x, *args):
-            return x
-
-        curve = curve_fit(
-            func, x, self._y, p0=2 * len(x) * [1], sigma=self.covariance_matrix
-        )
-        print(curve)
-
-    @property
-    def _y(self):
-        return np.array(
-            [
-                value
-                for samples in self.samples_list
-                for value in samples.max_log_likelihood(as_instance=False)
-            ]
-        )
